requirements.txt
# Core ML
torch>=2.0.0
transformers==4.44.2
scikit-learn>=1.3.0
tiktoken>=0.12.0
sentencepiece>=0.2.0
protobuf>=3.20.0

# API client
openai>=1.30.0

# Config
pyyaml>=6.0

# Optional: accelerate for faster training
accelerate>=0.29.0
